tests: divide tests into suites
