+++ Tests/Reader/DownloadFilesTest.php
@@ -2,67 +2,17 @@
 
 namespace Keboola\InputMapping\Tests\Reader;
 
-use Keboola\Csv\CsvFile;
 use Keboola\InputMapping\Configuration\File\Manifest\Adapter;
 use Keboola\InputMapping\Exception\InputOperationException;
 use Keboola\InputMapping\Exception\InvalidInputException;
 use Keboola\InputMapping\Reader\Reader;
 use Keboola\StorageApi\Client;
 use Keboola\StorageApi\Options\FileUploadOptions;
-use Keboola\StorageApi\Options\ListFilesOptions;
-use Keboola\Temp\Temp;
 use Psr\Log\NullLogger;
-use Symfony\Component\Filesystem\Filesystem;
 use Symfony\Component\Finder\Finder;
 
-class DownloadFilesTest extends \PHPUnit_Framework_TestCase
+class DownloadFilesTest extends DownloadFilesTestAbstract
 {
-    /**
-     * @var Client
-     */
-    protected $client;
-
-    /**
-     * @var string
-     */
-    protected $tmpDir;
-
-    /**
-     * @var Temp
-     */
-    private $temp;
-
-    public function setUp()
-    {
-        // Create folders
-        $temp = new Temp('docker');
-        $temp->initRunFolder();
-        $this->temp = $temp;
-        $this->tmpDir = $temp->getTmpFolder();
-        $fs = new Filesystem();
-        $fs->mkdir($this->tmpDir . "/download");
-        $this->client = new Client(["token" => STORAGE_API_TOKEN, "url" => STORAGE_API_URL]);
-    }
-
-    public function tearDown()
-    {
-        // Delete local files
-        $finder = new Finder();
-        $fs = new Filesystem();
-        $fs->remove($finder->files()->in($this->tmpDir . "/download"));
-        $fs->remove($finder->files()->in($this->tmpDir));
-        $fs->remove($this->tmpDir . "/download");
-        $fs->remove($this->tmpDir);
-
-        // Delete file uploads
-        $options = new ListFilesOptions();
-        $options->setTags(["docker-bundle-test"]);
-        $files = $this->client->listFiles($options);
-        foreach ($files as $file) {
-            $this->client->deleteFile($file["id"]);
-        }
-    }
-
     public function testReadFiles()
     {
         $root = $this->tmpDir;
@@ -193,70 +143,6 @@ class DownloadFilesTest extends \PHPUnit_Framework_TestCase
         self::assertTrue(file_exists($root . "/download/" . $id4 . '_upload'));
         self::assertTrue(file_exists($root . "/download/" . $id5 . '_upload'));
         self::assertTrue(file_exists($root . "/download/" . $id6 . '_upload'));
-    }
-
-    public function testReadSlicedFile()
-    {
-        // Create bucket
-        if (!$this->client->bucketExists("in.c-docker-test-redshift")) {
-            $this->client->createBucket("docker-test-redshift", Client::STAGE_IN, "Docker Testsuite", "redshift");
-        }
-
-        // Create redshift table and export it to produce a sliced file
-        if (!$this->client->tableExists("in.c-docker-test-redshift.test_file")) {
-            $csv = new CsvFile($this->tmpDir . "/upload.csv");
-            $csv->writeRow(["Id", "Name"]);
-            $csv->writeRow(["test", "test"]);
-            $this->client->createTableAsync("in.c-docker-test-redshift", "test_file", $csv);
-        }
-        $table = $this->client->exportTableAsync('in.c-docker-test-redshift.test_file');
-        $fileId = $table['file']['id'];
-
-        $reader = new Reader($this->client, new NullLogger());
-        $configuration = [['query' => 'id: ' . $fileId]];
-
-        $dlDir = $this->tmpDir . "/download";
-        $reader->downloadFiles($configuration, $dlDir);
-
-        $fileName = $fileId . "_in.c-docker-test-redshift.test_file.csv";
-        self::assertEquals(
-            '"test","test"' . "\n",
-            file_get_contents($dlDir . "/" . $fileName . "/part.0")
-            . file_get_contents($dlDir . "/" . $fileName . "/part.1")
-        );
-
-        $manifestFile = $dlDir . "/" . $fileName . ".manifest";
-        self::assertFileExists($manifestFile);
-        $adapter = new Adapter();
-        $manifest = $adapter->readFromFile($manifestFile);
-        self::assertArrayHasKey('is_sliced', $manifest);
-        self::assertTrue($manifest['is_sliced']);
-    }
-
-    public function testReadFilesEmptySlices()
-    {
-        $fileUploadOptions = new FileUploadOptions();
-        $fileUploadOptions
-            ->setIsSliced(true)
-            ->setFileName('empty_file');
-        $uploadFileId = $this->client->uploadSlicedFile([], $fileUploadOptions);
-        sleep(2);
-
-        $reader = new Reader($this->client, new NullLogger());
-        $configuration = [
-            [
-                'query' => 'id:' . $uploadFileId,
-            ],
-        ];
-        $reader->downloadFiles($configuration, $this->temp->getTmpFolder() . DIRECTORY_SEPARATOR . 'download');
-
-        $adapter = new Adapter();
-        $manifest = $adapter->readFromFile(
-            $this->temp->getTmpFolder() . '/download/' . $uploadFileId . '_empty_file.manifest'
-        );
-        self::assertEquals($uploadFileId, $manifest['id']);
-        self::assertEquals('empty_file', $manifest['name']);
-        self::assertDirectoryExists($this->temp->getTmpFolder() . '/download/' . $uploadFileId . '_empty_file');
     }
 
     public function testReadFilesLimit()
