+++ Tests/Reader/DownloadFilesTestAbstract.php
@@ -1,0 +1,57 @@
+<?php
+
+namespace Keboola\InputMapping\Tests\Reader;
+
+use Keboola\InputMapping\Configuration\File\Manifest\Adapter;
+use Keboola\InputMapping\Exception\InvalidInputException;
+use Keboola\InputMapping\Reader\Reader;
+use Keboola\StorageApi\Client;
+use Keboola\StorageApi\Options\FileUploadOptions;
+use Keboola\StorageApi\Options\ListFilesOptions;
+use Keboola\Temp\Temp;
+use Psr\Log\NullLogger;
+use Symfony\Component\Filesystem\Filesystem;
+use Symfony\Component\Finder\Finder;
+
+class DownloadFilesTestAbstract extends \PHPUnit_Framework_TestCase
+{
+    /** @var Client */
+    protected $client;
+
+    /** @var string */
+    protected $tmpDir;
+
+    /** @var Temp */
+    protected $temp;
+
+    public function setUp()
+    {
+        // Create folders
+        $temp = new Temp('docker');
+        $temp->initRunFolder();
+        $this->temp = $temp;
+        $this->tmpDir = $temp->getTmpFolder();
+        $fs = new Filesystem();
+        $fs->mkdir($this->tmpDir . "/download");
+        $this->client = new Client(["token" => STORAGE_API_TOKEN, "url" => STORAGE_API_URL]);
+    }
+
+    public function tearDown()
+    {
+        // Delete local files
+        $finder = new Finder();
+        $fs = new Filesystem();
+        $fs->remove($finder->files()->in($this->tmpDir . "/download"));
+        $fs->remove($finder->files()->in($this->tmpDir));
+        $fs->remove($this->tmpDir . "/download");
+        $fs->remove($this->tmpDir);
+
+        // Delete file uploads
+        $options = new ListFilesOptions();
+        $options->setTags(["docker-bundle-test"]);
+        $files = $this->client->listFiles($options);
+        foreach ($files as $file) {
+            $this->client->deleteFile($file["id"]);
+        }
+    }
+}
