+++ Tests/Reader/DownloadTablesRedshiftTest.php
@@ -9,6 +9,7 @@ use Keboola\InputMapping\Reader\Reader;
 use Keboola\InputMapping\Reader\State\InputTableStateList;
 use Keboola\StorageApi\Client;
 use Keboola\StorageApi\ClientException;
+use Keboola\StorageApi\Options\FileUploadOptions;
 use Psr\Log\NullLogger;
 
 class DownloadTablesRedshiftTest extends DownloadTablesTestAbstract
@@ -71,5 +72,38 @@ class DownloadTablesRedshiftTest extends DownloadTablesTestAbstract
         $manifest = $adapter->readFromFile($this->temp->getTmpFolder() . "/download/test-redshift.csv.manifest");
         self::assertEquals("in.c-docker-test-redshift.test", $manifest["id"]);
         $this->assertS3info($manifest);
+    }
+
+    public function testReadTablesEmptySlices()
+    {
+        $fileUploadOptions = new FileUploadOptions();
+        $fileUploadOptions
+            ->setIsSliced(true)
+            ->setFileName('emptyfile');
+        $uploadFileId = $this->client->uploadSlicedFile([], $fileUploadOptions);
+        $columns = ['Id', 'Name'];
+        $headerCsvFile = new CsvFile($this->temp->getTmpFolder() . 'header.csv');
+        $headerCsvFile->writeRow($columns);
+        $this->client->createTableAsync('in.c-docker-test', 'empty', $headerCsvFile, []);
+
+        $options['columns'] = $columns;
+        $options['dataFileId'] = $uploadFileId;
+        $this->client->writeTableAsyncDirect('in.c-docker-test.empty', $options);
+
+        $reader = new Reader($this->client, new NullLogger());
+        $configuration = new InputTableOptionsList([
+            [
+                "source" => "in.c-docker-test.empty",
+                "destination" => "empty.csv",
+            ],
+        ]);
+
+        $reader->downloadTables($configuration, new InputTableStateList([]), $this->temp->getTmpFolder() . DIRECTORY_SEPARATOR . "download");
+        $file = file_get_contents($this->temp->getTmpFolder() . "/download/empty.csv");
+        self::assertEquals("\"Id\",\"Name\"\n", $file);
+
+        $adapter = new Adapter();
+        $manifest = $adapter->readFromFile($this->temp->getTmpFolder() . "/download/empty.csv.manifest");
+        self::assertEquals("in.c-docker-test.empty", $manifest["id"]);
     }
 }
