+++ phpunit.xml.dist
@@ -10,9 +10,26 @@
          syntaxCheck="false"
          bootstrap="bootstrap.php">
 
-    <testsuite name="Test Suite">
-        <directory>Tests</directory>
-    </testsuite>
+    <testsuites>
+        <testsuite name="Common">
+            <directory>Tests/Configuration</directory>
+            <directory>Tests/Keboola</directory>
+            <directory>Tests/Reader/Options</directory>
+            <file>Tests/Reader/DownloadFilesTest.php</file>
+            <file>Tests/Reader/DownloadTablesAdaptiveTest.php</file>
+            <file>Tests/Reader/DownloadTablesDefaultTest.php</file>
+            <file>Tests/Reader/DownloadTablesOutputStateTest.php</file>
+            <file>Tests/Reader/ReaderTest.php</file>
+            <file>Tests/Reader/TableDefinitionResolverTest.php</file>
+        </testsuite>
+        <testsuite name="Redshift">
+            <file>Tests/Reader/DownloadFilesRedshiftTest.php</file>
+            <file>Tests/Reader/DownloadTablesRedshiftTest.php</file>
+        </testsuite>
+        <testsuite name="S3">
+            <file>Tests/Reader/DownloadTablesS3DefaultTest.php</file>
+        </testsuite>
+    </testsuites>
 
     <filter>
         <whitelist processUncoveredFilesFromWhitelist="true">
